Fix discord link display issue for ,ctan command in latexutil_cmds.py
When retrieving links from CTAN the link text if unspecified will be the link itself, meaning the markdown will be of the form `[https://link...](https://link...)`. Unfortunately Discord does not allow the string “`https://`” in the link text (presumably to prevent link spoofing), so this will display incorrectly. The fix is simple: just print the link by itself when the link text and link coincide.